tsoev/requirements.txt
nltk~=3.5
tqdm~=4.49.0
pandas~=1.2.4
PyMySQL~=1.0.2
Flask~=2.0.1
moviepy~=1.0.3
SQLAlchemy~=1.4.7
alembic~=1.5.8
spacy~=3.0.6
keybert~=0.3.0
simpletransformers~=0.61.5
numpy==1.20.3
requests==2.25.1
Flask-MySQL==1.5.2
